+++ CHANGELOG.md
@@ -7,10 +7,16 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
-# [0.0.2] - 2020-12-04
+## [0.0.3] - not released yet
 
 ### Added
-- new flag --alert-manager-target-alertname to add label prometheus_targets_url.
+- new flags `--sensu-extra-annotation string` and `--sensu-extra-label string` to add annotations and labels in each alert create in sensu agent api.
+
+## [0.0.2] - 2020-12-04
+
+### Added
+- new flag `--alert-manager-target-alertname` to add label prometheus_targets_url.
+
 
 ## [0.0.1] - 2020-11-30